Sync

+++ modules/plugin/imagemosaic/src/main/java/org/geotools/gce/imagemosaic/ImageMosaicReader.java
@@ -2,7 +2,7 @@
  *    GeoTools - The Open Source Java GIS Toolkit
  *    http://geotools.org
  *
- *    (C) 2006 - 2016, Open Source Geospatial Foundation (OSGeo)5
+ *    (C) 2006 - 2016, Open Source Geospatial Foundation (OSGeo)
  *
  *    This library is free software; you can redistribute it and/or
  *    modify it under the terms of the GNU Lesser General Public
@@ -16,6 +16,7 @@
  */
 package org.geotools.gce.imagemosaic;
 
+import it.geosolutions.imageio.maskband.DatasetLayout;
 import java.awt.Rectangle;
 import java.io.File;
 import java.io.FilenameFilter;
@@ -615,12 +616,14 @@ public class ImageMosaicReader extends AbstractGridCoverage2DReader
 
                 // Scan for MosaicConfigurationBeans from properties files
                 List<MosaicConfigurationBean> beans = new ArrayList<>();
-                for (File propFile : properties) {
-                    if (Utils.checkFileReadable(propFile)
-                            && Utils.loadMosaicProperties(URLs.fileToUrl(propFile)) != null) {
-                        configuration = Utils.loadMosaicProperties(URLs.fileToUrl(propFile));
-                        if (configuration != null) {
-                            beans.add(configuration);
+                if (properties != null) {
+                    for (File propFile : properties) {
+                        if (Utils.checkFileReadable(propFile)
+                                && Utils.loadMosaicProperties(URLs.fileToUrl(propFile)) != null) {
+                            configuration = Utils.loadMosaicProperties(URLs.fileToUrl(propFile));
+                            if (configuration != null) {
+                                beans.add(configuration);
+                            }
                         }
                     }
                 }
@@ -665,11 +668,6 @@ public class ImageMosaicReader extends AbstractGridCoverage2DReader
                     granuleCatalog.dispose();
                 }
                 granuleCatalog = catalog;
-
-                if (granuleCatalog == null) {
-                    throw new DataSourceException(
-                            "Unable to create index for this URL " + sourceURL);
-                }
 
                 // Creating a RasterManager for each mosaic configuration found on disk
                 for (MosaicConfigurationBean bean : beans) {
@@ -813,8 +811,8 @@ public class ImageMosaicReader extends AbstractGridCoverage2DReader
         if (suggestedSPIClass != null) {
             try {
                 final Class<?> clazz = Class.forName(suggestedSPIClass);
-                if (clazz.newInstance() instanceof ImageReaderSpi)
-                    suggestedSPI = (ImageReaderSpi) clazz.newInstance();
+                if (clazz.getDeclaredConstructor().newInstance() instanceof ImageReaderSpi)
+                    suggestedSPI = (ImageReaderSpi) clazz.getDeclaredConstructor().newInstance();
                 else suggestedSPI = null;
             } catch (Exception e) {
                 if (LOGGER.isLoggable(Level.FINE))
@@ -1074,28 +1072,20 @@ public class ImageMosaicReader extends AbstractGridCoverage2DReader
     }
 
     @Override
-    public int getNumOverviews(String coverageName) {
-        coverageName = checkUnspecifiedCoverage(coverageName);
-        RasterManager manager = getRasterManager(coverageName);
-        return manager.overviewsController.getNumberOfOverviews();
-    }
-
-    @Override
-    public int getNumOverviews() {
-        return getNumOverviews(UNSPECIFIED);
-    }
-
-    @Override
-    public double[] getReadingResolutions(OverviewPolicy policy, double[] requestedResolution) {
+    public double[] getReadingResolutions(OverviewPolicy policy, double[] requestedResolution)
+            throws IOException {
         return getReadingResolutions(UNSPECIFIED, policy, requestedResolution);
     }
 
     @Override
     public double[] getReadingResolutions(
-            String coverageName, OverviewPolicy policy, double[] requestedResolution) {
+            String coverageName, OverviewPolicy policy, double[] requestedResolution)
+            throws IOException {
         coverageName = checkUnspecifiedCoverage(coverageName);
         RasterManager manager = getRasterManager(coverageName);
-        final int numOverviews = getNumOverviews(coverageName);
+        DatasetLayout datasetLayout = getDatasetLayout(coverageName);
+        final int numOverviews =
+                datasetLayout.getNumInternalOverviews() + datasetLayout.getNumExternalOverviews();
         OverviewsController overviewsController = manager.overviewsController;
         OverviewLevel level = null;
         if (numOverviews > 0) {
@@ -1144,6 +1134,15 @@ public class ImageMosaicReader extends AbstractGridCoverage2DReader
             } else {
                 return coverageName;
             }
+        }
+    }
+
+    @Override
+    protected boolean checkName(String coverageName) {
+        if (coverageName.equalsIgnoreCase(UNSPECIFIED)) {
+            return getGridCoverageCount() == 1;
+        } else {
+            return names.contains(coverageName);
         }
     }
 
@@ -1374,6 +1373,9 @@ public class ImageMosaicReader extends AbstractGridCoverage2DReader
     @Override
     public List<DimensionDescriptor> getDimensionDescriptors(String coverageName)
             throws IOException {
+        if (coverageName == null) {
+            coverageName = defaultName;
+        }
         RasterManager manager = getRasterManager(coverageName);
         return manager.getDimensionDescriptors();
     }
@@ -1396,8 +1398,10 @@ public class ImageMosaicReader extends AbstractGridCoverage2DReader
         if (deleteData) {
             // quick way: delete everything
             final File[] list = parentDirectory.listFiles();
-            for (File file : list) {
-                FileUtils.deleteQuietly(file);
+            if (list != null) {
+                for (File file : list) {
+                    FileUtils.deleteQuietly(file);
+                }
             }
         } else {
             finalizeCleanup();
@@ -1415,12 +1419,6 @@ public class ImageMosaicReader extends AbstractGridCoverage2DReader
         for (File file : files) {
             FileUtils.deleteQuietly(file);
         }
-    }
-
-    @Override
-    public boolean removeCoverage(String coverageName)
-            throws IOException, UnsupportedOperationException {
-        return removeCoverage(coverageName, false);
     }
 
     /**
@@ -1540,5 +1538,10 @@ public class ImageMosaicReader extends AbstractGridCoverage2DReader
 
     public ExecutorService getMultiThreadedLoader() {
         return multiThreadedLoader;
+    }
+
+    public DatasetLayout getDatasetLayout() {
+        // Default implementation for backwards compatibility
+        return getDatasetLayout(checkUnspecifiedCoverage(UNSPECIFIED));
     }
 }
